Add files via upload
hello, i have fixed the exploding barrels, the tank ( which is meant to come down the screen detect player and then stop and fire) is still few issues please could you look at it at the momment it looks detects the player comes down the screen but however it never stops no matter what i get turn gravity to 0 or velocity nothing, i think it's something to do with the background moving down because when i added you're test background script to it and make scroll speed - 10 it stooped but did not move at all.  this also has a issue with the ai spawn i've added a few line of code in the ai health scrip but it does not seem to work if you could look at that as well please thank you.